Modified chat screen and MVVM structure
