+++ app/src/main/java/com/example/musical/MainActivity.kt
@@ -3,14 +3,6 @@ package com.example.musical
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.example.musical.ui.theme.MainView
 import com.example.musical.ui.theme.MusicalTheme
 
 class MainActivity : ComponentActivity() {
