+++ app/src/main/java/com/example/musical/common/utils/Reports.kt
@@ -1,8 +1,7 @@
-package com.example.musical.ui.theme
+package com.example.musical.common.utils
 
 import android.content.ActivityNotFoundException
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -23,7 +22,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
-import com.example.musical.R
 import android.content.Context
 import android.content.Intent
 import android.graphics.pdf.PdfDocument
@@ -42,6 +40,7 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
+import com.example.musical.R
 
 
 @Composable
