+++ app/src/main/java/com/example/musical/presentation/accounts/LogOut.kt
@@ -1,25 +1,15 @@
-package com.example.musical.ui.theme
+package com.example.musical.presentation.accounts
 
 import android.app.Activity
 import android.content.Context
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Card
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
 import androidx.compose.material.Text
-import androidx.compose.material.TextButton
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
-import androidx.compose.material.icons.filled.AccountBox
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -36,4 +26,9 @@ fun LogOut(context: Context){
             }
         )
     }
+}
+@Composable
+fun LogOut() {
+    val context = LocalContext.current
+    com.example.musical.presentation.accounts.LogOut(context = context)
 }
