+++ app/src/main/java/com/example/musical/presentation/home/components/BottomSheet.kt
@@ -1,0 +1,57 @@
+package com.example.musical.presentation.home.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.musical.R
+
+
+@Composable
+fun MoreBottomSheet(modifier: Modifier) {
+    Box(
+        Modifier
+            .fillMaxWidth()
+            .height(300.dp)
+            .background(Color(0xFF006eff))
+    ) {
+        Column(modifier = modifier.padding(16.dp), verticalArrangement = Arrangement.SpaceBetween) {
+            Row(modifier = modifier.padding(16.dp)) {
+                Icon(
+                    modifier = Modifier.padding(end = 8.dp),
+                    painter = painterResource(id = R.drawable.baseline_settings_24),
+                    contentDescription = "Settings"
+                )
+                Text(text = "Settings", fontSize = 20.sp, color = Color.White)
+            }
+            Row(modifier = modifier.padding(16.dp)) {
+                Icon(
+                    modifier = Modifier.padding(end = 8.dp),
+                    painter = painterResource(id = R.drawable.ic_baseline_share_24),
+                    contentDescription = "Share"
+                )
+                Text(text = "Share", fontSize = 20.sp, color = Color.White)
+            }
+            Row(modifier = modifier.padding(16.dp)) {
+                Icon(
+                    modifier = Modifier.padding(end = 8.dp),
+                    painter = painterResource(id = R.drawable.ic_help_green),
+                    contentDescription = "Help"
+                )
+                Text(text = "Help", fontSize = 20.sp, color = Color.White)
+            }
+        }
+    }
+}
