--- app/src/main/java/com/example/musical/ui/theme/HomeView.kt
@@ -1,601 +1,0 @@
-package com.example.musical.ui.theme
-
-
-import androidx.compose.foundation.Canvas
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ElevatedCard
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.*
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import com.example.musical.R
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Chat
-import androidx.compose.material.icons.filled.LocalHospital
-import androidx.compose.material.icons.filled.MedicalServices
-import androidx.compose.material.icons.filled.Medication
-import androidx.compose.material.icons.filled.HealthAndSafety
-import androidx.compose.material.icons.filled.Search
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.text.font.FontWeight
-import androidx.navigation.NavController
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.ui.unit.sp
-import androidx.compose.foundation.border
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.window.Dialog
-import com.google.accompanist.pager.*
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.yield
-
-
-
-data class Appointment(
-    val doctorName: String,
-    val specialization: String,
-    val imageRes: Int
-)
-
-
-
-@Composable
-fun Home(navController: NavController) {
-    var searchQuery by remember { mutableStateOf("") }
-    var showChatDialog by remember { mutableStateOf(false) }
-
-    Scaffold(
-        floatingActionButton = {
-            FloatingActionButton(
-                onClick = { showChatDialog = true },
-                containerColor = MaterialTheme.colorScheme.primary
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.Chat,
-                    contentDescription = "Chat"
-                )
-            }
-        },
-        content = { padding ->
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxSize()
-                    .background(Color(0xFFF5F5F5))
-                    .padding(padding)
-                    .padding(top = 8.dp),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Top
-            ) {
-                item {
-                    ImageCarousel()
-                }
-
-                item {
-                    Spacer(modifier = Modifier.height(16.dp))
-                }
-
-                item {
-                    SearchBar(searchQuery = searchQuery, onQueryChange = { searchQuery = it })
-                }
-
-                item {
-                    Spacer(modifier = Modifier.height(16.dp))
-                }
-
-                item {
-                    AppointmentsCarousel(navController = navController)
-                }
-
-                item {
-                    Spacer(modifier = Modifier.height(16.dp))
-                }
-
-                item {
-                    Text(
-                        text = "Statistics",
-                        style = MaterialTheme.typography.titleLarge,
-                        modifier = Modifier.padding(16.dp)
-                    )
-                }
-
-                // Example data for the charts
-                val barData = listOf(50f, 100f, 150f, 200f, 250f)
-                val pieData = listOf(30f, 40f, 50f, 80f)
-
-                item {
-                    ElevatedCard(
-                        elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier.padding(16.dp),
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            Text(
-                                text = "Bar Graph",
-                                style = MaterialTheme.typography.titleLarge
-                            )
-                            Spacer(modifier = Modifier.height(16.dp))
-                            BarGraph(
-                                data = barData,
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(200.dp)
-                            )
-                        }
-                    }
-                }
-
-                item {
-                    Spacer(modifier = Modifier.height(16.dp))
-                }
-
-                item {
-                    ElevatedCard(
-                        elevation = CardDefaults.cardElevation(defaultElevation = 6.dp),
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(16.dp)
-                    ) {
-                        Column(
-                            modifier = Modifier.padding(16.dp),
-                            horizontalAlignment = Alignment.CenterHorizontally
-                        ) {
-                            Text(
-                                text = "Pie Chart",
-                                style = MaterialTheme.typography.titleLarge
-                            )
-                            Spacer(modifier = Modifier.height(16.dp))
-                            PieChart(
-                                data = pieData,
-                                modifier = Modifier
-                                    .fillMaxWidth()
-                                    .height(200.dp)
-                            )
-                        }
-                    }
-                }
-            }
-
-            if (showChatDialog) {
-                ChatDialog(onDismiss = { showChatDialog = false })
-            }
-        }
-    )
-}
-
-
-
-
-@OptIn(ExperimentalPagerApi::class)
-@Composable
-fun AppointmentsCarousel(navController: NavController) {
-    val pagerState = rememberPagerState()
-    val appointments = listOf(
-        Appointment("Dr. Daksh Dhaka", "Heart Surgeon", R.drawable.your_image1),
-        Appointment("Dr. Jane Smith", "Pediatrician", R.drawable.your_image2),
-        Appointment("Dr. John Doe", "Dermatologist", R.drawable.your_image3),
-        Appointment("Dr. Alice Brown", "Neurologist", R.drawable.your_image4)
-    )
-
-    LaunchedEffect(pagerState) {
-        while (true) {
-            yield()
-            delay(3000) // 3 seconds delay
-            pagerState.animateScrollToPage((pagerState.currentPage + 1) % appointments.size)
-        }
-    }
-
-    HorizontalPager(
-        count = appointments.size,
-        state = pagerState,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(200.dp)
-    ) { page ->
-        AppointmentCard(navController = navController, appointment = appointments[page])
-    }
-}
-
-@Composable
-fun AppointmentCard(navController: NavController, appointment: Appointment) {
-    ElevatedCard(
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = 6.dp
-        ),
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(150.dp)
-            .padding(horizontal = 16.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color(0xFF4894fe))
-                .padding(16.dp)
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.Start,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxSize()
-            ) {
-                Box(
-                    modifier = Modifier
-                        .size(80.dp)
-                        .clip(CircleShape)
-                ) {
-                    Image(
-                        painter = painterResource(id = appointment.imageRes),
-                        contentDescription = "Doctor Image",
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .clip(CircleShape)
-                    )
-                }
-                Spacer(modifier = Modifier.width(16.dp))
-                Column {
-                    Text(
-                        text = appointment.doctorName,
-                        style = MaterialTheme.typography.titleLarge.copy(
-                            fontWeight = FontWeight.Bold,
-                            color = Color(0xFFFFFFFF)
-                        )
-                    )
-                    Text(
-                        text = appointment.specialization,
-                        style = MaterialTheme.typography.bodyMedium.copy(
-                            color = Color(0xFFFFFFFF)
-                        )
-                    )
-                }
-            }
-
-            Row(
-                modifier = Modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(16.dp)
-                    .clickable {
-                        navController.navigate("patient_details")
-                    }
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.Add,
-                    contentDescription = "Book Appointment",
-                    tint = Color.White,
-                    modifier = Modifier
-                        .size(24.dp)
-                        .background(Color(0xFF094999), CircleShape)
-                        .padding(4.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = "Book Appointment",
-                    style = MaterialTheme.typography.bodyMedium.copy(
-                        color = Color(0xFFFFFFFF)
-                    ),
-                    fontSize = 14.sp
-                )
-            }
-        }
-    }
-}
-
-
-@Composable
-fun ChatDialog(onDismiss: () -> Unit) {
-    var userInput by remember { mutableStateOf("") }
-    val chatMessages = remember { mutableStateListOf("How are you?") }
-
-    fun sendMessage(message: String) {
-        if (message.isNotBlank()) {
-            chatMessages.add(message)
-            // Simulate bot response
-            chatMessages.add("I am a bot, I received your message: $message")
-            userInput = ""
-        }
-    }
-
-    Dialog(onDismissRequest = { onDismiss() }) {
-        Surface(
-            shape = RoundedCornerShape(16.dp),
-            color = Color.White,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            Column(
-                modifier = Modifier
-                    .padding(16.dp)
-                    .background(Color.White),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Text(
-                    text = "Chat with Bot",
-                    style = MaterialTheme.typography.titleLarge,
-                    color = Color.Black
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                Column(
-                    modifier = Modifier
-                        .weight(1f)
-                        .fillMaxWidth()
-                        .padding(8.dp)
-                        .background(Color(0xFFF0F0F0))
-                        .clip(RoundedCornerShape(8.dp))
-                        .padding(8.dp)
-                ) {
-                    chatMessages.forEach { message ->
-                        Text(
-                            text = message,
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = Color.Black,
-                            modifier = Modifier.padding(4.dp)
-                        )
-                    }
-                }
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 8.dp),
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.SpaceBetween
-                ) {
-                    OutlinedTextField(
-                        value = userInput,
-                        onValueChange = { userInput = it },
-                        placeholder = { Text(text = "Type your message...") },
-                        modifier = Modifier.weight(1f)
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Button(onClick = { sendMessage(userInput) }) {
-                        Text(text = "Send")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun BarGraph(data: List<Float>, modifier: Modifier = Modifier) {
-    val maxValue = data.maxOrNull() ?: 0f
-    Canvas(modifier = modifier) {
-        val barWidth = size.width / data.size
-        data.forEachIndexed { index, value ->
-            val barHeight = (value / maxValue) * size.height
-            drawRect(
-                color = Color(0xFF4894fe),
-                topLeft = Offset(x = index * barWidth, y = size.height - barHeight),
-                size = Size(width = barWidth, height = barHeight)
-            )
-        }
-    }
-}
-
-@Composable
-fun PieChart(data: List<Float>, modifier: Modifier = Modifier) {
-    val totalValue = data.sum()
-    var startAngle = 0f
-    Canvas(modifier = modifier) {
-        data.forEach { value ->
-            val sweepAngle = (value / totalValue) * 360f
-            drawArc(
-                color = Color(0xFF4894fe),
-                startAngle = startAngle,
-                sweepAngle = sweepAngle,
-                useCenter = true,
-                topLeft = Offset.Zero,
-                size = size
-            )
-            startAngle += sweepAngle
-        }
-    }
-}
-
-
-@Composable
-fun AppointmentCard(navController: NavController) {
-    ElevatedCard(
-        elevation = CardDefaults.cardElevation(
-            defaultElevation = 6.dp
-        ),
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(150.dp)
-            .padding(horizontal = 16.dp)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxSize()
-                .background(Color(0xFF4894fe))
-                .padding(16.dp)
-        ) {
-            Row(
-                horizontalArrangement = Arrangement.Start,
-                verticalAlignment = Alignment.CenterVertically,
-                modifier = Modifier
-                    .fillMaxSize()
-            ) {
-                Box(
-                    modifier = Modifier
-                        .size(80.dp)
-                        .clip(CircleShape)
-                ) {
-                    Image(
-                        painter = painterResource(id = R.drawable.your_image1), // Replace with your image resource
-                        contentDescription = "Doctor Image",
-                        modifier = Modifier
-                            .fillMaxSize()
-                            .clip(CircleShape)
-                    )
-                }
-                Spacer(modifier = Modifier.width(16.dp))
-                Column {
-                    Text(
-                        text = "Dr. Daksh Dhaka",
-                        style = MaterialTheme.typography.titleLarge.copy(
-                            fontWeight = FontWeight.Bold,
-                            color = Color(0xFFFFFFFF)
-                        )
-                    )
-                    Text(
-                        text = "Heart Surgeon",
-                        style = MaterialTheme.typography.bodyMedium.copy(
-                            color = Color(0xFFFFFFFF)
-                        )
-                    )
-                }
-            }
-
-            Row(
-                modifier = Modifier
-                    .align(Alignment.BottomEnd)
-                    .padding(16.dp)
-                    .clickable {
-                        navController.navigate("patient_details")
-                    }
-            ) {
-                Icon(
-                    imageVector = Icons.Filled.Add,
-                    contentDescription = "Book Appointment",
-                    tint = Color.White,
-                    modifier = Modifier
-                        .size(24.dp)
-                        .background(Color(0xFF094999), CircleShape)
-                        .padding(4.dp)
-                )
-                Spacer(modifier = Modifier.width(8.dp))
-                Text(
-                    text = "Book Appointment",
-                    style = MaterialTheme.typography.bodyMedium.copy(
-                        color = Color(0xFFFFFFFF)
-                    ),
-                    fontSize = 14.sp
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun SearchBar(searchQuery: String, onQueryChange: (String) -> Unit) {
-    OutlinedTextField(
-        value = searchQuery,
-        onValueChange = onQueryChange,
-        label = { Text("Search") },
-        leadingIcon = {
-            Icon(
-                imageVector = Icons.Filled.Search,
-                contentDescription = "Search Icon"
-            )
-        },
-        shape = RoundedCornerShape(24.dp), // Adjust the value for desired roundness
-        colors = TextFieldDefaults.outlinedTextFieldColors(
-            focusedBorderColor = MaterialTheme.colorScheme.primary,
-            unfocusedBorderColor = Color.Gray,
-            containerColor = MaterialTheme.colorScheme.surface,
-            cursorColor = MaterialTheme.colorScheme.primary, // Cursor color
-            focusedLabelColor = MaterialTheme.colorScheme.primary,
-            unfocusedLabelColor = Color.Gray
-        ),
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp)
-            .background(
-                color = Color.White.copy(alpha = 0.5f),
-                shape = RoundedCornerShape(24.dp)
-            )
-            .shadow(
-                elevation = 8.dp,
-                shape = RoundedCornerShape(24.dp),
-                clip = true
-            )
-            .padding(4.dp)
-    )
-}
-
-
-
-@OptIn(ExperimentalPagerApi::class)
-@Composable
-fun ImageCarousel() {
-    val pagerState = rememberPagerState()
-    val images = listOf(
-        Pair(R.drawable.img1, "Hospital 1 - New York"),
-        Pair(R.drawable.img2, "Hospital 2 - Los Angeles"),
-        Pair(R.drawable.img3, "Hospital 3 - Chicago"),
-        Pair(R.drawable.img4, "Hospital 4 - Houston")
-    )
-    LaunchedEffect(pagerState) {
-        while (true) {
-            yield()
-            delay(3000) // 3 seconds delay
-            pagerState.animateScrollToPage((pagerState.currentPage + 1) % images.size)
-        }
-    }
-    HorizontalPager(
-        count = images.size,
-        state = pagerState,
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(200.dp)
-    ) { page ->
-        Card(
-            shape = RoundedCornerShape(16.dp),
-            modifier = Modifier
-                .padding(16.dp)
-                .fillMaxSize()
-                .clickable { /* Handle click action */ },
-            elevation = CardDefaults.cardElevation(
-                defaultElevation = 6.dp
-            ),
-        ) {
-            Box {
-                Image(
-                    painter = painterResource(id = images[page].first),
-                    contentDescription = "Image ${page + 1}",
-                    modifier = Modifier
-                        .fillMaxSize(),
-                    contentScale = ContentScale.Crop
-                )
-                Text(
-                    text = images[page].second,
-                    color = Color.White,
-                    style = MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.Bold),
-                    modifier = Modifier
-                        .align(Alignment.BottomStart)
-                        .background(Color.Black.copy(alpha = 0.5f))
-                        .padding(8.dp)
-                )
-            }
-        }
-    }
-}
-
-
-
